chore: deploy task uses DeployCommandHandler
